doctorq-api/pyproject.toml
[project]
name = "doctorq-api"
version = "0.1.0"
description = "DoctorQ API - Plataforma SaaS para gestão de clínicas de estética com IA, inspirada no Doctoralia."
authors = [{ name = "Equipe DoctorQ", email = "devs@doctorq.app" }]
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn[standard]>=0.34.3,<0.35.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "pydantic[email]>=2.11.7,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "PyJWT>=2.9.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "asyncpg>=0.30.0",
    "psutil>=7.1.0",
    "langchain-core>=0.3.78",
    "colorlog>=6.9.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.6.0",
    "langchain-openai>=0.3.35",
    "cryptography>=46.0.2",
    "langchain>=0.3.27",
    "redis>=6.4.0",
    "pgvector>=0.4.1",
    "langchain-community>=0.3.30",
    "langfuse>=3.6.1",
    "docling>=2.55.1",
    "qdrant-client>=1.15.1",
    "msal>=1.31.1",
    "python-multipart>=0.0.6",
    "gunicorn>=23.0.0",
    "langchain-classic>=1.0.0",
    "pillow>=11.0.0",
    "piexif>=1.1.3",
    "stripe>=11.3.0",
    "mercadopago>=2.2.4",
    "bcrypt>=4.0.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pylint>=3.0.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "alembic>=1.13.1",
]


[tool.black]
line-length = 88
target-version = ["py312"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pylint]
disable = [
    "W0707"
]

[tool.pyright]
reportAttributeAccessIssue = "none"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
